pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.48",
    "openai>=2.2.0",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "pymupdf>=1.26.4",
    "pypdf>=6.1.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
